requirements.txt
yfinance==0.2.65
pandas==2.3.1
numpy==2.3.1
colorama==0.4.6
tabulate==0.9.0
requests==2.32.4
python-telegram-bot==21.9
